Moved meta_enc and meta_raw into meta module
