+++ src/encryption.rs
@@ -1,7 +1,7 @@
 use crate::error;
 use crate::meta::error::MetaError;
 use crate::meta::header::MetaHeader;
-use crate::meta_raw::RawMeta;
+use crate::meta::raw::RawMeta;
 use chacha20poly1305::{
     aead::{stream, NewAead},
     XChaCha20Poly1305,
