+++ src/error/mod.rs
@@ -65,7 +65,7 @@ impl ErrorKind {
             FileAlreadyExist => "File already exist",
             WrongPassword => "Wrong password",
             FileTooSmall => "File is too small",
-            FileMetaDecodeError => "File's meta_raw decode error",
+            FileMetaDecodeError => "File's meta decode error",
             FileInvalidMagic => "Invalid file magic",
             FileInvalidCipherId => "Invalid file cipher ID",
             EncryptedMetaIsEmpty => "Encrypted meta is empty",
