+++ src/meta/mod.rs
@@ -1,2 +1,4 @@
+pub mod enc;
 pub mod error;
 pub mod header;
+pub mod raw;
